Add files via upload
The GUI work is still in progress (design and colour are not given; I will add them after checking and testing the app's functionality). Also, the format in which the folder and log.txt should be selected is given inside the help.html page. This is a desktop application, currently creating a webpage using python-flask, where the user can download the exe-file.